1. 다음은 클래스 분리 - CustomListModel
2. state 사용하기
